requirements.txt
flask
mysql-connector
reportlab
Flask-Compress